backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Pillow==10.0.0
opencv-python==4.8.1.78
numpy==1.24.3
scipy==1.11.2
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==3.0.0
